requirements.txt
fastapi
uvicorn
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic
